+++ src/utils/stringUtils.ts
